backend/requirements.txt
fastembed
numpy
pypdf
django
django-allauth
dj-rest-auth[with_social]
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-celery-results
django-extensions
psycopg2-binary
pgvector
channels
channels-redis
daphne
celery
redis
groq
httpx
speechmatics-rt>=0.5.1
pyaudio>=0.2.11
python-dotenv>=1.0.0
requests
whitenoise
gunicorn
